requirements.txt
apache-airflow
apache-airflow-providers-http
kafka-python
requests
pandas
xgboost
PyWavelets
numpy
psycopg2-binary
SQLAlchemy
joblib
scikit-learn
scipy
statsmodels
nolds